moving util to member function temporarily
