+++ planning/behavior_velocity_intersection_module/src/scene_intersection.cpp
@@ -17,7 +17,10 @@
 #include "util.hpp"
 
 #include <behavior_velocity_planner_common/utilization/boost_geometry_helper.hpp>
+#include <behavior_velocity_planner_common/utilization/trajectory_utils.hpp>
 #include <behavior_velocity_planner_common/utilization/util.hpp>
+#include <lanelet2_extension/regulatory_elements/road_marking.hpp>
+#include <lanelet2_extension/utility/query.hpp>
 #include <lanelet2_extension/utility/utilities.hpp>
 #include <motion_utils/trajectory/trajectory.hpp>
 #include <opencv2/imgproc.hpp>
@@ -37,10 +40,12 @@
 
 #include <algorithm>
 #include <limits>
+#include <list>
 #include <map>
 #include <memory>
 #include <tuple>
 #include <vector>
+
 namespace behavior_velocity_planner
 {
 namespace bg = boost::geometry;
@@ -70,6 +75,25 @@ Polygon2d createOneStepPolygon(
   return convex_one_step_poly;
 }
 }  // namespace
+
+static bool isTargetStuckVehicleType(
+  const autoware_auto_perception_msgs::msg::PredictedObject & object)
+{
+  if (
+    object.classification.at(0).label ==
+      autoware_auto_perception_msgs::msg::ObjectClassification::CAR ||
+    object.classification.at(0).label ==
+      autoware_auto_perception_msgs::msg::ObjectClassification::BUS ||
+    object.classification.at(0).label ==
+      autoware_auto_perception_msgs::msg::ObjectClassification::TRUCK ||
+    object.classification.at(0).label ==
+      autoware_auto_perception_msgs::msg::ObjectClassification::TRAILER ||
+    object.classification.at(0).label ==
+      autoware_auto_perception_msgs::msg::ObjectClassification::MOTORCYCLE) {
+    return true;
+  }
+  return false;
+}
 
 static bool isTargetCollisionVehicleType(
   const autoware_auto_perception_msgs::msg::PredictedObject & object)
@@ -376,7 +400,7 @@ void reactRTCApprovalByDecisionResult(
   const bool rtc_default_approved, const bool rtc_occlusion_approved, const T & decision_result,
   const IntersectionModule::PlannerParam & planner_param, const double baselink2front,
   autoware_auto_planning_msgs::msg::PathWithLaneId * path, StopReason * stop_reason,
-  VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   static_assert("Unsupported type passed to reactRTCByDecisionResult");
   return;
@@ -392,7 +416,7 @@ void reactRTCApprovalByDecisionResult(
   [[maybe_unused]] autoware_auto_planning_msgs::msg::PathWithLaneId * path,
   [[maybe_unused]] StopReason * stop_reason,
   [[maybe_unused]] VelocityFactorInterface * velocity_factor,
-  [[maybe_unused]] util::DebugData * debug_data)
+  [[maybe_unused]] DebugData * debug_data)
 {
   return;
 }
@@ -403,7 +427,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::StuckStop & decision_result,
   [[maybe_unused]] const IntersectionModule::PlannerParam & planner_param,
   const double baselink2front, autoware_auto_planning_msgs::msg::PathWithLaneId * path,
-  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -451,7 +475,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::YieldStuckStop & decision_result,
   [[maybe_unused]] const IntersectionModule::PlannerParam & planner_param,
   const double baselink2front, autoware_auto_planning_msgs::msg::PathWithLaneId * path,
-  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -483,7 +507,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::NonOccludedCollisionStop & decision_result,
   [[maybe_unused]] const IntersectionModule::PlannerParam & planner_param,
   const double baselink2front, autoware_auto_planning_msgs::msg::PathWithLaneId * path,
-  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -526,7 +550,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::FirstWaitBeforeOcclusion & decision_result,
   const IntersectionModule::PlannerParam & planner_param, const double baselink2front,
   autoware_auto_planning_msgs::msg::PathWithLaneId * path, StopReason * stop_reason,
-  VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -577,7 +601,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::PeekingTowardOcclusion & decision_result,
   const IntersectionModule::PlannerParam & planner_param, const double baselink2front,
   autoware_auto_planning_msgs::msg::PathWithLaneId * path, StopReason * stop_reason,
-  VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -634,7 +658,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::OccludedCollisionStop & decision_result,
   [[maybe_unused]] const IntersectionModule::PlannerParam & planner_param,
   const double baselink2front, autoware_auto_planning_msgs::msg::PathWithLaneId * path,
-  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -681,7 +705,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::OccludedAbsenceTrafficLight & decision_result,
   [[maybe_unused]] const IntersectionModule::PlannerParam & planner_param,
   const double baselink2front, autoware_auto_planning_msgs::msg::PathWithLaneId * path,
-  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -734,7 +758,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::Safe & decision_result,
   [[maybe_unused]] const IntersectionModule::PlannerParam & planner_param,
   const double baselink2front, autoware_auto_planning_msgs::msg::PathWithLaneId * path,
-  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -776,7 +800,7 @@ void reactRTCApprovalByDecisionResult(
   const IntersectionModule::FullyPrioritized & decision_result,
   [[maybe_unused]] const IntersectionModule::PlannerParam & planner_param,
   const double baselink2front, autoware_auto_planning_msgs::msg::PathWithLaneId * path,
-  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  StopReason * stop_reason, VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   RCLCPP_DEBUG(
     rclcpp::get_logger("reactRTCApprovalByDecisionResult"),
@@ -818,7 +842,7 @@ void reactRTCApproval(
   const IntersectionModule::DecisionResult & decision_result,
   const IntersectionModule::PlannerParam & planner_param, const double baselink2front,
   autoware_auto_planning_msgs::msg::PathWithLaneId * path, StopReason * stop_reason,
-  VelocityFactorInterface * velocity_factor, util::DebugData * debug_data)
+  VelocityFactorInterface * velocity_factor, DebugData * debug_data)
 {
   std::visit(
     VisitorSwitch{[&](const auto & decision) {
@@ -868,7 +892,7 @@ static std::string formatDecisionResult(const IntersectionModule::DecisionResult
 
 bool IntersectionModule::modifyPathVelocity(PathWithLaneId * path, StopReason * stop_reason)
 {
-  debug_data_ = util::DebugData();
+  debug_data_ = DebugData();
   *stop_reason = planning_utils::initializeStopReason(StopReason::INTERSECTION);
 
   // set default RTC
@@ -956,7 +980,7 @@ IntersectionModule::DecisionResult IntersectionModule::modifyPathVelocityDetail(
   const auto lanelets_on_path =
     planning_utils::getLaneletsOnPath(*path, lanelet_map_ptr, current_pose);
   if (!intersection_lanelets_) {
-    intersection_lanelets_ = util::getObjectiveLanelets(
+    intersection_lanelets_ = getObjectiveLanelets(
       lanelet_map_ptr, routing_graph_ptr, assigned_lanelet, lanelets_on_path, associative_ids_,
       planner_param_.common.attention_area_length,
       planner_param_.occlusion.occlusion_attention_area_length,
@@ -964,12 +988,12 @@ IntersectionModule::DecisionResult IntersectionModule::modifyPathVelocityDetail(
   }
   auto & intersection_lanelets = intersection_lanelets_.value();
 
-  // at the very first time of registration of this module, the path may not be conflicting with the
+  // at the very first time of regisTration of this module, the path may not be conflicting with the
   // attention area, so update() is called to update the internal data as well as traffic light info
   const auto traffic_prioritized_level =
-    util::getTrafficPrioritizedLevel(assigned_lanelet, planner_data_->traffic_light_id_map);
+    getTrafficPrioritizedLevel(assigned_lanelet, planner_data_->traffic_light_id_map);
   const bool is_prioritized =
-    traffic_prioritized_level == util::TrafficPrioritizedLevel::FULLY_PRIORITIZED;
+    traffic_prioritized_level == TrafficPrioritizedLevel::FULLY_PRIORITIZED;
   const auto footprint = planner_data_->vehicle_info_.createFootprint(0.0, 0.0);
   intersection_lanelets.update(is_prioritized, interpolated_path_info, footprint, baselink2front);
 
@@ -994,7 +1018,7 @@ IntersectionModule::DecisionResult IntersectionModule::modifyPathVelocityDetail(
     intersection_lanelets.first_attention_lane()
       ? intersection_lanelets.first_attention_lane().value().centerline2d()
       : first_conflicting_lane.centerline2d();
-  const auto intersection_stoplines_opt = util::generateIntersectionStopLines(
+  const auto intersection_stoplines_opt = generateIntersectionStopLines(
     first_conflicting_area, dummy_first_attention_area, dummy_first_attention_lane_centerline,
     planner_data_, interpolated_path_info, planner_param_.stuck_vehicle.use_stuck_stopline,
     planner_param_.common.default_stopline_margin, planner_param_.common.max_accel,
@@ -1011,7 +1035,7 @@ IntersectionModule::DecisionResult IntersectionModule::modifyPathVelocityDetail(
 
   // see the doc for struct PathLanelets
   const auto & conflicting_area = intersection_lanelets.conflicting_area();
-  const auto path_lanelets_opt = util::generatePathLanelets(
+  const auto path_lanelets_opt = generatePathLanelets(
     lanelets_on_path, interpolated_path_info, associative_ids_, first_conflicting_area,
     conflicting_area, first_attention_area_opt, intersection_lanelets.attention_area(), closest_idx,
     planner_data_->vehicle_info_.vehicle_width_m);
@@ -1157,8 +1181,8 @@ IntersectionModule::DecisionResult IntersectionModule::modifyPathVelocityDetail(
   }
 
   const double is_amber_or_red =
-    (traffic_prioritized_level == util::TrafficPrioritizedLevel::PARTIALLY_PRIORITIZED) ||
-    (traffic_prioritized_level == util::TrafficPrioritizedLevel::FULLY_PRIORITIZED);
+    (traffic_prioritized_level == TrafficPrioritizedLevel::PARTIALLY_PRIORITIZED) ||
+    (traffic_prioritized_level == TrafficPrioritizedLevel::FULLY_PRIORITIZED);
   auto occlusion_status =
     (enable_occlusion_detection_ && !occlusion_attention_lanelets.empty() && !is_amber_or_red)
       ? getOcclusionStatus(
@@ -1384,7 +1408,7 @@ IntersectionModule::DecisionResult IntersectionModule::modifyPathVelocityDetail(
 }
 
 bool IntersectionModule::checkStuckVehicle(
-  const std::shared_ptr<const PlannerData> & planner_data, const util::PathLanelets & path_lanelets)
+  const std::shared_ptr<const PlannerData> & planner_data, const PathLanelets & path_lanelets)
 {
   const bool stuck_detection_direction = [&]() {
     return (turn_direction_ == "left" && planner_param_.stuck_vehicle.turn_direction.left) ||
@@ -1398,18 +1422,17 @@ bool IntersectionModule::checkStuckVehicle(
   const auto & objects_ptr = planner_data->predicted_objects;
 
   // considering lane change in the intersection, these lanelets are generated from the path
-  const auto stuck_vehicle_detect_area = util::generateStuckVehicleDetectAreaPolygon(
+  const auto stuck_vehicle_detect_area = generateStuckVehicleDetectAreaPolygon(
     path_lanelets, planner_param_.stuck_vehicle.stuck_vehicle_detect_dist);
   debug_data_.stuck_vehicle_detect_area = toGeomPoly(stuck_vehicle_detect_area);
 
-  return util::checkStuckVehicleInIntersection(
+  return checkStuckVehicleInIntersection(
     objects_ptr, stuck_vehicle_detect_area,
     planner_param_.stuck_vehicle.stuck_vehicle_velocity_threshold, &debug_data_);
 }
 
 bool IntersectionModule::checkYieldStuckVehicle(
-  const util::TargetObjects & target_objects,
-  const util::InterpolatedPathInfo & interpolated_path_info,
+  const TargetObjects & target_objects, const util::InterpolatedPathInfo & interpolated_path_info,
   const lanelet::ConstLanelets & attention_lanelets)
 {
   const bool yield_stuck_detection_direction = [&]() {
@@ -1421,20 +1444,20 @@ bool IntersectionModule::checkYieldStuckVehicle(
     return false;
   }
 
-  return util::checkYieldStuckVehicleInIntersection(
+  return checkYieldStuckVehicleInIntersection(
     target_objects, interpolated_path_info, attention_lanelets, turn_direction_,
     planner_data_->vehicle_info_.vehicle_width_m,
     planner_param_.stuck_vehicle.stuck_vehicle_velocity_threshold,
     planner_param_.yield_stuck.distance_threshold, &debug_data_);
 }
 
-util::TargetObjects IntersectionModule::generateTargetObjects(
-  const util::IntersectionLanelets & intersection_lanelets,
+TargetObjects IntersectionModule::generateTargetObjects(
+  const IntersectionLanelets & intersection_lanelets,
   const std::optional<Polygon2d> & intersection_area) const
 {
   const auto & objects_ptr = planner_data_->predicted_objects;
   // extract target objects
-  util::TargetObjects target_objects;
+  TargetObjects target_objects;
   target_objects.header = objects_ptr->header;
   const auto & attention_lanelets = intersection_lanelets.attention();
   const auto & attention_lanelet_stoplines = intersection_lanelets.attention_stoplines();
@@ -1447,7 +1470,7 @@ util::TargetObjects IntersectionModule::generateTargetObjects(
 
     // check direction of objects
     const auto object_direction = util::getObjectPoseWithVelocityDirection(object.kinematics);
-    const auto belong_adjacent_lanelet_id = util::checkAngleForTargetLanelets(
+    const auto belong_adjacent_lanelet_id = checkAngleForTargetLanelets(
       object_direction, adjacent_lanelets, planner_param_.common.attention_area_angle_threshold,
       planner_param_.collision_detection.consider_wrong_direction_vehicle,
       planner_param_.common.attention_area_margin, false);
@@ -1464,25 +1487,25 @@ util::TargetObjects IntersectionModule::generateTargetObjects(
       const auto & obj_pos = object.kinematics.initial_pose_with_covariance.pose.position;
       const auto obj_poly = tier4_autoware_utils::toPolygon2d(object);
       const auto intersection_area_2d = intersection_area.value();
-      const auto belong_attention_lanelet_id = util::checkAngleForTargetLanelets(
+      const auto belong_attention_lanelet_id = checkAngleForTargetLanelets(
         object_direction, attention_lanelets, planner_param_.common.attention_area_angle_threshold,
         planner_param_.collision_detection.consider_wrong_direction_vehicle,
         planner_param_.common.attention_area_margin, is_parked_vehicle);
       if (belong_attention_lanelet_id) {
         const auto id = belong_attention_lanelet_id.value();
-        util::TargetObject target_object;
+        TargetObject target_object;
         target_object.object = object;
         target_object.attention_lanelet = attention_lanelets.at(id);
         target_object.stopline = attention_lanelet_stoplines.at(id);
         container.push_back(target_object);
       } else if (bg::within(Point2d{obj_pos.x, obj_pos.y}, intersection_area_2d)) {
-        util::TargetObject target_object;
+        TargetObject target_object;
         target_object.object = object;
         target_object.attention_lanelet = std::nullopt;
         target_object.stopline = std::nullopt;
         target_objects.intersection_area_objects.push_back(target_object);
       }
-    } else if (const auto belong_attention_lanelet_id = util::checkAngleForTargetLanelets(
+    } else if (const auto belong_attention_lanelet_id = checkAngleForTargetLanelets(
                  object_direction, attention_lanelets,
                  planner_param_.common.attention_area_angle_threshold,
                  planner_param_.collision_detection.consider_wrong_direction_vehicle,
@@ -1490,7 +1513,7 @@ util::TargetObjects IntersectionModule::generateTargetObjects(
                belong_attention_lanelet_id.has_value()) {
       // intersection_area is not available, use detection_area_with_margin as before
       const auto id = belong_attention_lanelet_id.value();
-      util::TargetObject target_object;
+      TargetObject target_object;
       target_object.object = object;
       target_object.attention_lanelet = attention_lanelets.at(id);
       target_object.stopline = attention_lanelet_stoplines.at(id);
@@ -1510,10 +1533,10 @@ util::TargetObjects IntersectionModule::generateTargetObjects(
 }
 
 bool IntersectionModule::checkCollision(
-  const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
-  util::TargetObjects * target_objects, const util::PathLanelets & path_lanelets,
-  const size_t closest_idx, const size_t last_intersection_stopline_candidate_idx,
-  const double time_delay, const util::TrafficPrioritizedLevel & traffic_prioritized_level)
+  const autoware_auto_planning_msgs::msg::PathWithLaneId & path, TargetObjects * target_objects,
+  const PathLanelets & path_lanelets, const size_t closest_idx,
+  const size_t last_intersection_stopline_candidate_idx, const double time_delay,
+  const TrafficPrioritizedLevel & traffic_prioritized_level)
 {
   using lanelet::utils::getArcCoordinates;
   using lanelet::utils::getPolygonFromArcLength;
@@ -1521,7 +1544,7 @@ bool IntersectionModule::checkCollision(
   // check collision between target_objects predicted path and ego lane
   // cut the predicted path at passing_time
   tier4_debug_msgs::msg::Float64MultiArrayStamped ego_ttc_time_array;
-  const auto time_distance_array = util::calcIntersectionPassingTime(
+  const auto time_distance_array = calcIntersectionPassingTime(
     path, planner_data_, lane_id_, associative_ids_, closest_idx,
     last_intersection_stopline_candidate_idx, time_delay,
     planner_param_.collision_detection.velocity_profile.default_velocity,
@@ -1538,7 +1561,7 @@ bool IntersectionModule::checkCollision(
   }
 
   const double passing_time = time_distance_array.back().first;
-  util::cutPredictPathWithDuration(target_objects, clock_, passing_time);
+  cutPredictPathWithDuration(target_objects, clock_, passing_time);
 
   const auto & concat_lanelets = path_lanelets.all;
   const auto closest_arc_coords = getArcCoordinates(
@@ -1549,12 +1572,12 @@ bool IntersectionModule::checkCollision(
 
   // change TTC margin based on ego traffic light color
   const auto [collision_start_margin_time, collision_end_margin_time] = [&]() {
-    if (traffic_prioritized_level == util::TrafficPrioritizedLevel::FULLY_PRIORITIZED) {
+    if (traffic_prioritized_level == TrafficPrioritizedLevel::FULLY_PRIORITIZED) {
       return std::make_pair(
         planner_param_.collision_detection.fully_prioritized.collision_start_margin_time,
         planner_param_.collision_detection.fully_prioritized.collision_end_margin_time);
     }
-    if (traffic_prioritized_level == util::TrafficPrioritizedLevel::PARTIALLY_PRIORITIZED) {
+    if (traffic_prioritized_level == TrafficPrioritizedLevel::PARTIALLY_PRIORITIZED) {
       return std::make_pair(
         planner_param_.collision_detection.partially_prioritized.collision_start_margin_time,
         planner_param_.collision_detection.partially_prioritized.collision_end_margin_time);
@@ -1563,7 +1586,7 @@ bool IntersectionModule::checkCollision(
       planner_param_.collision_detection.not_prioritized.collision_start_margin_time,
       planner_param_.collision_detection.not_prioritized.collision_end_margin_time);
   }();
-  const auto expectedToStopBeforeStopLine = [&](const util::TargetObject & target_object) {
+  const auto expectedToStopBeforeStopLine = [&](const TargetObject & target_object) {
     if (!target_object.dist_to_stopline) {
       return false;
     }
@@ -1578,7 +1601,7 @@ bool IntersectionModule::checkCollision(
                          .object_expected_deceleration));
     return dist_to_stopline > braking_distance;
   };
-  const auto isTolerableOvershoot = [&](const util::TargetObject & target_object) {
+  const auto isTolerableOvershoot = [&](const TargetObject & target_object) {
     if (
       !target_object.attention_lanelet || !target_object.dist_to_stopline ||
       !target_object.stopline) {
@@ -1637,14 +1660,14 @@ bool IntersectionModule::checkCollision(
     // If the vehicle is expected to stop before their stopline, ignore
     const bool expected_to_stop_before_stopline = expectedToStopBeforeStopLine(target_object);
     if (
-      traffic_prioritized_level == util::TrafficPrioritizedLevel::PARTIALLY_PRIORITIZED &&
+      traffic_prioritized_level == TrafficPrioritizedLevel::PARTIALLY_PRIORITIZED &&
       expected_to_stop_before_stopline) {
       debug_data_.amber_ignore_targets.objects.push_back(object);
       continue;
     }
     const bool is_tolerable_overshoot = isTolerableOvershoot(target_object);
     if (
-      traffic_prioritized_level == util::TrafficPrioritizedLevel::FULLY_PRIORITIZED &&
+      traffic_prioritized_level == TrafficPrioritizedLevel::FULLY_PRIORITIZED &&
       !expected_to_stop_before_stopline && is_tolerable_overshoot) {
       debug_data_.red_overshoot_ignore_targets.objects.push_back(object);
       continue;
@@ -1771,7 +1794,7 @@ IntersectionModule::OcclusionType IntersectionModule::getOcclusionStatus(
   const lanelet::CompoundPolygon3d & first_attention_area,
   const util::InterpolatedPathInfo & interpolated_path_info,
   const std::vector<lanelet::ConstLineString3d> & lane_divisions,
-  const util::TargetObjects & target_objects, const geometry_msgs::msg::Pose & current_pose,
+  const TargetObjects & target_objects, const geometry_msgs::msg::Pose & current_pose,
   const double occlusion_dist_thr)
 {
   const auto & path_ip = interpolated_path_info.path;
@@ -2090,6 +2113,1196 @@ IntersectionModule::OcclusionType IntersectionModule::getOcclusionStatus(
     }
   }
   return OcclusionType::STATICALLY_OCCLUDED;
+}
+
+// utilから移動
+Polygon2d IntersectionModule::generateStuckVehicleDetectAreaPolygon(
+  const PathLanelets & path_lanelets, const double stuck_vehicle_detect_dist)
+{
+  using lanelet::utils::getArcCoordinates;
+  using lanelet::utils::getLaneletLength3d;
+  using lanelet::utils::getPolygonFromArcLength;
+  using lanelet::utils::to2D;
+
+  Polygon2d polygon{};
+  if (path_lanelets.conflicting_interval_and_remaining.size() == 0) {
+    return polygon;
+  }
+
+  double target_polygon_length =
+    getLaneletLength3d(path_lanelets.conflicting_interval_and_remaining);
+  lanelet::ConstLanelets targets = path_lanelets.conflicting_interval_and_remaining;
+  if (path_lanelets.next) {
+    targets.push_back(path_lanelets.next.value());
+    const double next_arc_length =
+      std::min(stuck_vehicle_detect_dist, getLaneletLength3d(path_lanelets.next.value()));
+    target_polygon_length += next_arc_length;
+  }
+  const auto target_polygon =
+    to2D(getPolygonFromArcLength(targets, 0, target_polygon_length)).basicPolygon();
+
+  if (target_polygon.empty()) {
+    return polygon;
+  }
+
+  for (const auto & p : target_polygon) {
+    polygon.outer().emplace_back(p.x(), p.y());
+  }
+
+  polygon.outer().emplace_back(polygon.outer().front());
+  bg::correct(polygon);
+
+  return polygon;
+}
+
+std::optional<size_t> IntersectionModule::checkAngleForTargetLanelets(
+  const geometry_msgs::msg::Pose & pose, const lanelet::ConstLanelets & target_lanelets,
+  const double detection_area_angle_thr, const bool consider_wrong_direction_vehicle,
+  const double dist_margin, const bool is_parked_vehicle) const
+{
+  for (unsigned i = 0; i < target_lanelets.size(); ++i) {
+    const auto & ll = target_lanelets.at(i);
+    if (!lanelet::utils::isInLanelet(pose, ll, dist_margin)) {
+      continue;
+    }
+    const double ll_angle = lanelet::utils::getLaneletAngle(ll, pose.position);
+    const double pose_angle = tf2::getYaw(pose.orientation);
+    const double angle_diff = tier4_autoware_utils::normalizeRadian(ll_angle - pose_angle, -M_PI);
+    if (consider_wrong_direction_vehicle) {
+      if (std::fabs(angle_diff) > 1.57 || std::fabs(angle_diff) < detection_area_angle_thr) {
+        return std::make_optional<size_t>(i);
+      }
+    } else {
+      if (std::fabs(angle_diff) < detection_area_angle_thr) {
+        return std::make_optional<size_t>(i);
+      }
+      // NOTE: sometimes parked vehicle direction is reversed even if its longitudinal velocity is
+      // positive
+      if (
+        is_parked_vehicle && (std::fabs(angle_diff) < detection_area_angle_thr ||
+                              (std::fabs(angle_diff + M_PI) < detection_area_angle_thr))) {
+        return std::make_optional<size_t>(i);
+      }
+    }
+  }
+  return std::nullopt;
+}
+
+void IntersectionModule::cutPredictPathWithDuration(
+  TargetObjects * target_objects, const rclcpp::Clock::SharedPtr clock, const double time_thr)
+{
+  const rclcpp::Time current_time = clock->now();
+  for (auto & target_object : target_objects->all_attention_objects) {  // each objects
+    for (auto & predicted_path :
+         target_object.object.kinematics.predicted_paths) {  // each predicted paths
+      const auto origin_path = predicted_path;
+      predicted_path.path.clear();
+
+      for (size_t k = 0; k < origin_path.path.size(); ++k) {  // each path points
+        const auto & predicted_pose = origin_path.path.at(k);
+        const auto predicted_time =
+          rclcpp::Time(target_objects->header.stamp) +
+          rclcpp::Duration(origin_path.time_step) * static_cast<double>(k);
+        if ((predicted_time - current_time).seconds() < time_thr) {
+          predicted_path.path.push_back(predicted_pose);
+        }
+      }
+    }
+  }
+}
+
+TimeDistanceArray IntersectionModule::calcIntersectionPassingTime(
+  const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
+  const std::shared_ptr<const PlannerData> & planner_data, const lanelet::Id lane_id,
+  const std::set<lanelet::Id> & associative_ids, const size_t closest_idx,
+  const size_t last_intersection_stopline_candidate_idx, const double time_delay,
+  const double intersection_velocity, const double minimum_ego_velocity,
+  const bool use_upstream_velocity, const double minimum_upstream_velocity,
+  tier4_debug_msgs::msg::Float64MultiArrayStamped * debug_ttc_array)
+{
+  const double current_velocity = planner_data->current_velocity->twist.linear.x;
+
+  int assigned_lane_found = false;
+
+  // crop intersection part of the path, and set the reference velocity to intersection_velocity
+  // for ego's ttc
+  PathWithLaneId reference_path;
+  std::optional<size_t> upstream_stopline{std::nullopt};
+  for (size_t i = 0; i < path.points.size() - 1; ++i) {
+    auto reference_point = path.points.at(i);
+    // assume backward velocity is current ego velocity
+    if (i < closest_idx) {
+      reference_point.point.longitudinal_velocity_mps = current_velocity;
+    }
+    if (
+      i > last_intersection_stopline_candidate_idx &&
+      std::fabs(reference_point.point.longitudinal_velocity_mps) <
+        std::numeric_limits<double>::epsilon() &&
+      !upstream_stopline) {
+      upstream_stopline = i;
+    }
+    if (!use_upstream_velocity) {
+      reference_point.point.longitudinal_velocity_mps = intersection_velocity;
+    }
+    reference_path.points.push_back(reference_point);
+    bool has_objective_lane_id = util::hasLaneIds(path.points.at(i), associative_ids);
+    if (assigned_lane_found && !has_objective_lane_id) {
+      break;
+    }
+    assigned_lane_found = has_objective_lane_id;
+  }
+  if (!assigned_lane_found) {
+    return {{0.0, 0.0}};  // has already passed the intersection.
+  }
+
+  std::vector<std::pair<double, double>> original_path_xy;
+  for (size_t i = 0; i < reference_path.points.size(); ++i) {
+    const auto & p = reference_path.points.at(i).point.pose.position;
+    original_path_xy.emplace_back(p.x, p.y);
+  }
+
+  // apply smoother to reference velocity
+  PathWithLaneId smoothed_reference_path = reference_path;
+  if (!smoothPath(reference_path, smoothed_reference_path, planner_data)) {
+    smoothed_reference_path = reference_path;
+  }
+
+  // calculate when ego is going to reach each (interpolated) points on the path
+  TimeDistanceArray time_distance_array{};
+  double dist_sum = 0.0;
+  double passing_time = time_delay;
+  time_distance_array.emplace_back(passing_time, dist_sum);
+
+  // NOTE: `reference_path` is resampled in `reference_smoothed_path`, so
+  // `last_intersection_stopline_candidate_idx` makes no sense
+  const auto smoothed_path_closest_idx = motion_utils::findFirstNearestIndexWithSoftConstraints(
+    smoothed_reference_path.points, path.points.at(closest_idx).point.pose,
+    planner_data->ego_nearest_dist_threshold, planner_data->ego_nearest_yaw_threshold);
+
+  const std::optional<size_t> upstream_stopline_idx_opt = [&]() -> std::optional<size_t> {
+    if (upstream_stopline) {
+      const auto upstream_stopline_point = path.points.at(upstream_stopline.value()).point.pose;
+      return motion_utils::findFirstNearestIndexWithSoftConstraints(
+        smoothed_reference_path.points, upstream_stopline_point,
+        planner_data->ego_nearest_dist_threshold, planner_data->ego_nearest_yaw_threshold);
+    } else {
+      return std::nullopt;
+    }
+  }();
+  const bool has_upstream_stopline = upstream_stopline_idx_opt.has_value();
+  const size_t upstream_stopline_ind = upstream_stopline_idx_opt.value_or(0);
+
+  for (size_t i = smoothed_path_closest_idx; i < smoothed_reference_path.points.size() - 1; ++i) {
+    const auto & p1 = smoothed_reference_path.points.at(i);
+    const auto & p2 = smoothed_reference_path.points.at(i + 1);
+
+    const double dist = tier4_autoware_utils::calcDistance2d(p1, p2);
+    dist_sum += dist;
+
+    // use average velocity between p1 and p2
+    const double average_velocity =
+      (p1.point.longitudinal_velocity_mps + p2.point.longitudinal_velocity_mps) / 2.0;
+    const double passing_velocity = [=]() {
+      if (use_upstream_velocity) {
+        if (has_upstream_stopline && i > upstream_stopline_ind) {
+          return minimum_upstream_velocity;
+        }
+        return std::max<double>(average_velocity, minimum_ego_velocity);
+      } else {
+        return std::max<double>(average_velocity, minimum_ego_velocity);
+      }
+    }();
+    passing_time += (dist / passing_velocity);
+
+    time_distance_array.emplace_back(passing_time, dist_sum);
+  }
+  debug_ttc_array->layout.dim.resize(3);
+  debug_ttc_array->layout.dim.at(0).label = "lane_id_@[0][0], ttc_time, ttc_dist, path_x, path_y";
+  debug_ttc_array->layout.dim.at(0).size = 5;
+  debug_ttc_array->layout.dim.at(1).label = "values";
+  debug_ttc_array->layout.dim.at(1).size = time_distance_array.size();
+  debug_ttc_array->data.reserve(
+    time_distance_array.size() * debug_ttc_array->layout.dim.at(0).size);
+  for (unsigned i = 0; i < time_distance_array.size(); ++i) {
+    debug_ttc_array->data.push_back(lane_id);
+  }
+  for (const auto & [t, d] : time_distance_array) {
+    debug_ttc_array->data.push_back(t);
+  }
+  for (const auto & [t, d] : time_distance_array) {
+    debug_ttc_array->data.push_back(d);
+  }
+  for (size_t i = smoothed_path_closest_idx; i < smoothed_reference_path.points.size(); ++i) {
+    const auto & p = smoothed_reference_path.points.at(i).point.pose.position;
+    debug_ttc_array->data.push_back(p.x);
+  }
+  for (size_t i = smoothed_path_closest_idx; i < smoothed_reference_path.points.size(); ++i) {
+    const auto & p = smoothed_reference_path.points.at(i).point.pose.position;
+    debug_ttc_array->data.push_back(p.y);
+  }
+  return time_distance_array;
+}
+
+/*
+double calcDistanceUntilIntersectionLanelet(
+  const lanelet::ConstLanelet & assigned_lanelet,
+  const autoware_auto_planning_msgs::msg::PathWithLaneId & path, const size_t closest_idx)
+{
+  const auto lane_id = assigned_lanelet.id();
+  const auto intersection_first_itr =
+    std::find_if(path.points.cbegin(), path.points.cend(), [&](const auto & p) {
+      return std::find(p.lane_ids.begin(), p.lane_ids.end(), lane_id) != p.lane_ids.end();
+    });
+  if (
+    intersection_first_itr == path.points.begin() || intersection_first_itr == path.points.end()) {
+    return 0.0;
+  }
+  const auto dst_idx = std::distance(path.points.begin(), intersection_first_itr) - 1;
+
+  if (closest_idx > static_cast<size_t>(dst_idx)) {
+    return 0.0;
+  }
+
+  double distance = std::abs(motion_utils::calcSignedArcLength(path.points, closest_idx, dst_idx));
+  const auto & lane_first_point = assigned_lanelet.centerline2d().front();
+  distance += std::hypot(
+    path.points.at(dst_idx).point.pose.position.x - lane_first_point.x(),
+    path.points.at(dst_idx).point.pose.position.y - lane_first_point.y());
+  return distance;
+}
+*/
+
+static std::optional<std::pair<
+  size_t /* the index of interpolated PathPoint*/, size_t /* the index of corresponding Polygon */>>
+getFirstPointInsidePolygonsByFootprint(
+  const std::vector<lanelet::CompoundPolygon3d> & polygons,
+  const util::InterpolatedPathInfo & interpolated_path_info,
+  const tier4_autoware_utils::LinearRing2d & footprint, const double vehicle_length)
+{
+  const auto & path_ip = interpolated_path_info.path;
+  const auto [lane_start, lane_end] = interpolated_path_info.lane_id_interval.value();
+  const size_t vehicle_length_idx = static_cast<size_t>(vehicle_length / interpolated_path_info.ds);
+  const size_t start =
+    static_cast<size_t>(std::max<int>(0, static_cast<int>(lane_start) - vehicle_length_idx));
+
+  for (size_t i = start; i <= lane_end; ++i) {
+    const auto & pose = path_ip.points.at(i).point.pose;
+    const auto path_footprint =
+      tier4_autoware_utils::transformVector(footprint, tier4_autoware_utils::pose2transform(pose));
+    for (size_t j = 0; j < polygons.size(); ++j) {
+      const auto area_2d = lanelet::utils::to2D(polygons.at(j)).basicPolygon();
+      const bool is_in_polygon = bg::intersects(area_2d, path_footprint);
+      if (is_in_polygon) {
+        return std::make_optional<std::pair<size_t, size_t>>(i, j);
+      }
+    }
+  }
+  return std::nullopt;
+}
+
+void IntersectionLanelets::update(
+  const bool is_prioritized, const util::InterpolatedPathInfo & interpolated_path_info,
+  const tier4_autoware_utils::LinearRing2d & footprint, const double vehicle_length)
+{
+  is_prioritized_ = is_prioritized;
+  // find the first conflicting/detection area polygon intersecting the path
+  if (!first_conflicting_area_) {
+    auto first = getFirstPointInsidePolygonsByFootprint(
+      conflicting_area_, interpolated_path_info, footprint, vehicle_length);
+    if (first) {
+      first_conflicting_lane_ = conflicting_.at(first.value().second);
+      first_conflicting_area_ = conflicting_area_.at(first.value().second);
+    }
+  }
+  if (!first_attention_area_) {
+    auto first = getFirstPointInsidePolygonsByFootprint(
+      attention_non_preceding_area_, interpolated_path_info, footprint, vehicle_length);
+    if (first) {
+      first_attention_lane_ = attention_non_preceding_.at(first.value().second);
+      first_attention_area_ = attention_non_preceding_area_.at(first.value().second);
+    }
+  }
+}
+
+static lanelet::ConstLanelets getPrevLanelets(
+  const lanelet::ConstLanelets & lanelets_on_path, const std::set<lanelet::Id> & associative_ids)
+{
+  lanelet::ConstLanelets previous_lanelets;
+  for (const auto & ll : lanelets_on_path) {
+    if (associative_ids.find(ll.id()) != associative_ids.end()) {
+      return previous_lanelets;
+    }
+    previous_lanelets.push_back(ll);
+  }
+  return previous_lanelets;
+}
+
+// end inclusive
+static lanelet::ConstLanelet generatePathLanelet(
+  const PathWithLaneId & path, const size_t start_idx, const size_t end_idx, const double width,
+  const double interval)
+{
+  lanelet::Points3d lefts;
+  lanelet::Points3d rights;
+  size_t prev_idx = start_idx;
+  for (size_t i = start_idx; i <= end_idx; ++i) {
+    const auto & p = path.points.at(i).point.pose;
+    const auto & p_prev = path.points.at(prev_idx).point.pose;
+    if (i != start_idx && tier4_autoware_utils::calcDistance2d(p_prev, p) < interval) {
+      continue;
+    }
+    prev_idx = i;
+    const double yaw = tf2::getYaw(p.orientation);
+    const double x = p.position.x;
+    const double y = p.position.y;
+    // NOTE: maybe this is opposite
+    const double left_x = x + width / 2 * std::sin(yaw);
+    const double left_y = y - width / 2 * std::cos(yaw);
+    const double right_x = x - width / 2 * std::sin(yaw);
+    const double right_y = y + width / 2 * std::cos(yaw);
+    lefts.emplace_back(lanelet::InvalId, left_x, left_y, p.position.z);
+    rights.emplace_back(lanelet::InvalId, right_x, right_y, p.position.z);
+  }
+  lanelet::LineString3d left = lanelet::LineString3d(lanelet::InvalId, lefts);
+  lanelet::LineString3d right = lanelet::LineString3d(lanelet::InvalId, rights);
+
+  return lanelet::Lanelet(lanelet::InvalId, left, right);
+}
+
+static std::optional<std::pair<size_t, const lanelet::CompoundPolygon3d &>>
+getFirstPointInsidePolygons(
+  const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
+  const std::pair<size_t, size_t> lane_interval,
+  const std::vector<lanelet::CompoundPolygon3d> & polygons, const bool search_forward = true)
+{
+  if (search_forward) {
+    for (size_t i = lane_interval.first; i <= lane_interval.second; ++i) {
+      bool is_in_lanelet = false;
+      const auto & p = path.points.at(i).point.pose.position;
+      for (const auto & polygon : polygons) {
+        const auto polygon_2d = lanelet::utils::to2D(polygon);
+        is_in_lanelet = bg::within(to_bg2d(p), polygon_2d);
+        if (is_in_lanelet) {
+          return std::make_optional<std::pair<size_t, const lanelet::CompoundPolygon3d &>>(
+            i, polygon);
+        }
+      }
+      if (is_in_lanelet) {
+        break;
+      }
+    }
+  } else {
+    for (size_t i = lane_interval.second; i >= lane_interval.first; --i) {
+      bool is_in_lanelet = false;
+      const auto & p = path.points.at(i).point.pose.position;
+      for (const auto & polygon : polygons) {
+        const auto polygon_2d = lanelet::utils::to2D(polygon);
+        is_in_lanelet = bg::within(to_bg2d(p), polygon_2d);
+        if (is_in_lanelet) {
+          return std::make_optional<std::pair<size_t, const lanelet::CompoundPolygon3d &>>(
+            i, polygon);
+        }
+      }
+      if (is_in_lanelet) {
+        break;
+      }
+      if (i == 0) {
+        break;
+      }
+    }
+  }
+  return std::nullopt;
+}
+
+std::optional<PathLanelets> IntersectionModule::generatePathLanelets(
+  const lanelet::ConstLanelets & lanelets_on_path,
+  const util::InterpolatedPathInfo & interpolated_path_info,
+  const std::set<lanelet::Id> & associative_ids,
+  const lanelet::CompoundPolygon3d & first_conflicting_area,
+  const std::vector<lanelet::CompoundPolygon3d> & conflicting_areas,
+  const std::optional<lanelet::CompoundPolygon3d> & first_attention_area,
+  const std::vector<lanelet::CompoundPolygon3d> & attention_areas, const size_t closest_idx,
+  const double width)
+{
+  static constexpr double path_lanelet_interval = 1.5;
+
+  const auto & assigned_lane_interval_opt = interpolated_path_info.lane_id_interval;
+  if (!assigned_lane_interval_opt) {
+    return std::nullopt;
+  }
+  const auto assigned_lane_interval = assigned_lane_interval_opt.value();
+  const auto & path = interpolated_path_info.path;
+
+  PathLanelets path_lanelets;
+  // prev
+  path_lanelets.prev = getPrevLanelets(lanelets_on_path, associative_ids);
+  path_lanelets.all = path_lanelets.prev;
+
+  // entry2ego if exist
+  const auto [assigned_lane_start, assigned_lane_end] = assigned_lane_interval;
+  if (closest_idx > assigned_lane_start) {
+    path_lanelets.all.push_back(
+      generatePathLanelet(path, assigned_lane_start, closest_idx, width, path_lanelet_interval));
+  }
+
+  // ego_or_entry2exit
+  const auto ego_or_entry_start = std::max(closest_idx, assigned_lane_start);
+  path_lanelets.ego_or_entry2exit =
+    generatePathLanelet(path, ego_or_entry_start, assigned_lane_end, width, path_lanelet_interval);
+  path_lanelets.all.push_back(path_lanelets.ego_or_entry2exit);
+
+  // next
+  if (assigned_lane_end < path.points.size() - 1) {
+    const int next_id = path.points.at(assigned_lane_end).lane_ids.at(0);
+    const auto next_lane_interval_opt = util::findLaneIdsInterval(path, {next_id});
+    if (next_lane_interval_opt) {
+      const auto [next_start, next_end] = next_lane_interval_opt.value();
+      path_lanelets.next =
+        generatePathLanelet(path, next_start, next_end, width, path_lanelet_interval);
+      path_lanelets.all.push_back(path_lanelets.next.value());
+    }
+  }
+
+  const auto first_inside_conflicting_idx_opt =
+    first_attention_area.has_value()
+      ? util::getFirstPointInsidePolygon(path, assigned_lane_interval, first_attention_area.value())
+      : util::getFirstPointInsidePolygon(path, assigned_lane_interval, first_conflicting_area);
+  const auto last_inside_conflicting_idx_opt =
+    first_attention_area.has_value()
+      ? getFirstPointInsidePolygons(path, assigned_lane_interval, attention_areas, false)
+      : getFirstPointInsidePolygons(path, assigned_lane_interval, conflicting_areas, false);
+  if (first_inside_conflicting_idx_opt && last_inside_conflicting_idx_opt) {
+    const auto first_inside_conflicting_idx = first_inside_conflicting_idx_opt.value();
+    const auto last_inside_conflicting_idx = last_inside_conflicting_idx_opt.value().first;
+    lanelet::ConstLanelet conflicting_interval = generatePathLanelet(
+      path, first_inside_conflicting_idx, last_inside_conflicting_idx, width,
+      path_lanelet_interval);
+    path_lanelets.conflicting_interval_and_remaining.push_back(std::move(conflicting_interval));
+    if (last_inside_conflicting_idx < assigned_lane_end) {
+      lanelet::ConstLanelet remaining_interval = generatePathLanelet(
+        path, last_inside_conflicting_idx, assigned_lane_end, width, path_lanelet_interval);
+      path_lanelets.conflicting_interval_and_remaining.push_back(std::move(remaining_interval));
+    }
+  }
+  return path_lanelets;
+}
+
+void TargetObject::calc_dist_to_stopline()
+{
+  if (!attention_lanelet || !stopline) {
+    return;
+  }
+  const auto attention_lanelet_val = attention_lanelet.value();
+  const auto object_arc_coords = lanelet::utils::getArcCoordinates(
+    {attention_lanelet_val}, object.kinematics.initial_pose_with_covariance.pose);
+  const auto stopline_val = stopline.value();
+  geometry_msgs::msg::Pose stopline_center;
+  stopline_center.position.x = (stopline_val.front().x() + stopline_val.back().x()) / 2.0;
+  stopline_center.position.y = (stopline_val.front().y() + stopline_val.back().y()) / 2.0;
+  stopline_center.position.z = (stopline_val.front().z() + stopline_val.back().z()) / 2.0;
+  const auto stopline_arc_coords =
+    lanelet::utils::getArcCoordinates({attention_lanelet_val}, stopline_center);
+  dist_to_stopline = (stopline_arc_coords.length - object_arc_coords.length);
+}
+
+static lanelet::LineString3d getLineStringFromArcLength(
+  const lanelet::ConstLineString3d & linestring, const double s1, const double s2)
+{
+  lanelet::Points3d points;
+  double accumulated_length = 0;
+  size_t start_index = linestring.size();
+  for (size_t i = 0; i < linestring.size() - 1; i++) {
+    const auto & p1 = linestring[i];
+    const auto & p2 = linestring[i + 1];
+    const double length = boost::geometry::distance(p1.basicPoint(), p2.basicPoint());
+    if (accumulated_length + length > s1) {
+      start_index = i;
+      break;
+    }
+    accumulated_length += length;
+  }
+  if (start_index < linestring.size() - 1) {
+    const auto & p1 = linestring[start_index];
+    const auto & p2 = linestring[start_index + 1];
+    const double residue = s1 - accumulated_length;
+    const auto direction_vector = (p2.basicPoint() - p1.basicPoint()).normalized();
+    const auto start_basic_point = p1.basicPoint() + residue * direction_vector;
+    const auto start_point = lanelet::Point3d(lanelet::InvalId, start_basic_point);
+    points.push_back(start_point);
+  }
+
+  accumulated_length = 0;
+  size_t end_index = linestring.size();
+  for (size_t i = 0; i < linestring.size() - 1; i++) {
+    const auto & p1 = linestring[i];
+    const auto & p2 = linestring[i + 1];
+    const double length = boost::geometry::distance(p1.basicPoint(), p2.basicPoint());
+    if (accumulated_length + length > s2) {
+      end_index = i;
+      break;
+    }
+    accumulated_length += length;
+  }
+
+  for (size_t i = start_index + 1; i < end_index; i++) {
+    const auto p = lanelet::Point3d(linestring[i]);
+    points.push_back(p);
+  }
+  if (end_index < linestring.size() - 1) {
+    const auto & p1 = linestring[end_index];
+    const auto & p2 = linestring[end_index + 1];
+    const double residue = s2 - accumulated_length;
+    const auto direction_vector = (p2.basicPoint() - p1.basicPoint()).normalized();
+    const auto end_basic_point = p1.basicPoint() + residue * direction_vector;
+    const auto end_point = lanelet::Point3d(lanelet::InvalId, end_basic_point);
+    points.push_back(end_point);
+  }
+  return lanelet::LineString3d{lanelet::InvalId, points};
+}
+
+static lanelet::ConstLanelet createLaneletFromArcLength(
+  const lanelet::ConstLanelet & lanelet, const double s1, const double s2)
+{
+  const double total_length = boost::geometry::length(lanelet.centerline2d().basicLineString());
+  // make sure that s1, and s2 are between [0, lane_length]
+  const auto s1_saturated = std::max(0.0, std::min(s1, total_length));
+  const auto s2_saturated = std::max(0.0, std::min(s2, total_length));
+
+  const auto ratio_s1 = s1_saturated / total_length;
+  const auto ratio_s2 = s2_saturated / total_length;
+
+  const auto s1_left =
+    static_cast<double>(ratio_s1 * boost::geometry::length(lanelet.leftBound().basicLineString()));
+  const auto s2_left =
+    static_cast<double>(ratio_s2 * boost::geometry::length(lanelet.leftBound().basicLineString()));
+  const auto s1_right =
+    static_cast<double>(ratio_s1 * boost::geometry::length(lanelet.rightBound().basicLineString()));
+  const auto s2_right =
+    static_cast<double>(ratio_s2 * boost::geometry::length(lanelet.rightBound().basicLineString()));
+
+  const auto left_bound = getLineStringFromArcLength(lanelet.leftBound(), s1_left, s2_left);
+  const auto right_bound = getLineStringFromArcLength(lanelet.rightBound(), s1_right, s2_right);
+
+  return lanelet::Lanelet(lanelet::InvalId, left_bound, right_bound);
+}
+
+static std::vector<lanelet::CompoundPolygon3d> getPolygon3dFromLanelets(
+  const lanelet::ConstLanelets & ll_vec)
+{
+  std::vector<lanelet::CompoundPolygon3d> polys;
+  for (auto && ll : ll_vec) {
+    polys.push_back(ll.polygon3d());
+  }
+  return polys;
+}
+
+bool IntersectionModule::checkYieldStuckVehicleInIntersection(
+  const TargetObjects & target_objects, const util::InterpolatedPathInfo & interpolated_path_info,
+  const lanelet::ConstLanelets & attention_lanelets, const std::string & turn_direction,
+  const double width, const double stuck_vehicle_vel_thr, const double yield_stuck_distance_thr,
+  DebugData * debug_data)
+{
+  LineString2d sparse_intersection_path;
+  const auto [start, end] = interpolated_path_info.lane_id_interval.value();
+  for (unsigned i = start; i < end; ++i) {
+    const auto & point = interpolated_path_info.path.points.at(i).point.pose.position;
+    const auto yaw = tf2::getYaw(interpolated_path_info.path.points.at(i).point.pose.orientation);
+    if (turn_direction == "right") {
+      const double right_x = point.x - width / 2 * std::sin(yaw);
+      const double right_y = point.y + width / 2 * std::cos(yaw);
+      sparse_intersection_path.emplace_back(right_x, right_y);
+    } else if (turn_direction == "left") {
+      const double left_x = point.x + width / 2 * std::sin(yaw);
+      const double left_y = point.y - width / 2 * std::cos(yaw);
+      sparse_intersection_path.emplace_back(left_x, left_y);
+    } else {
+      // straight
+      sparse_intersection_path.emplace_back(point.x, point.y);
+    }
+  }
+  lanelet::ConstLanelets yield_stuck_detect_lanelets;
+  for (const auto & attention_lanelet : attention_lanelets) {
+    const auto centerline = attention_lanelet.centerline2d().basicLineString();
+    std::vector<Point2d> intersects;
+    bg::intersection(sparse_intersection_path, centerline, intersects);
+    if (intersects.empty()) {
+      continue;
+    }
+    const auto intersect = intersects.front();
+    const auto intersect_arc_coords = lanelet::geometry::toArcCoordinates(
+      centerline, lanelet::BasicPoint2d(intersect.x(), intersect.y()));
+    const double yield_stuck_start =
+      std::max(0.0, intersect_arc_coords.length - yield_stuck_distance_thr);
+    const double yield_stuck_end = intersect_arc_coords.length;
+    yield_stuck_detect_lanelets.push_back(
+      createLaneletFromArcLength(attention_lanelet, yield_stuck_start, yield_stuck_end));
+  }
+  debug_data->yield_stuck_detect_area = getPolygon3dFromLanelets(yield_stuck_detect_lanelets);
+  for (const auto & object : target_objects.all_attention_objects) {
+    const auto obj_v_norm = std::hypot(
+      object.object.kinematics.initial_twist_with_covariance.twist.linear.x,
+      object.object.kinematics.initial_twist_with_covariance.twist.linear.y);
+
+    if (obj_v_norm > stuck_vehicle_vel_thr) {
+      continue;
+    }
+    for (const auto & yield_stuck_detect_lanelet : yield_stuck_detect_lanelets) {
+      const bool is_in_lanelet = lanelet::utils::isInLanelet(
+        object.object.kinematics.initial_pose_with_covariance.pose, yield_stuck_detect_lanelet);
+      if (is_in_lanelet) {
+        debug_data->yield_stuck_targets.objects.push_back(object.object);
+        return true;
+      }
+    }
+  }
+  return false;
+}
+
+bool IntersectionModule::checkStuckVehicleInIntersection(
+  const autoware_auto_perception_msgs::msg::PredictedObjects::ConstSharedPtr objects_ptr,
+  const Polygon2d & stuck_vehicle_detect_area, const double stuck_vehicle_vel_thr,
+  DebugData * debug_data)
+{
+  for (const auto & object : objects_ptr->objects) {
+    if (!isTargetStuckVehicleType(object)) {
+      continue;  // not target vehicle type
+    }
+    const auto obj_v_norm = std::hypot(
+      object.kinematics.initial_twist_with_covariance.twist.linear.x,
+      object.kinematics.initial_twist_with_covariance.twist.linear.y);
+    if (obj_v_norm > stuck_vehicle_vel_thr) {
+      continue;  // not stop vehicle
+    }
+
+    // check if the footprint is in the stuck detect area
+    const auto obj_footprint = tier4_autoware_utils::toPolygon2d(object);
+    const bool is_in_stuck_area = !bg::disjoint(obj_footprint, stuck_vehicle_detect_area);
+    if (is_in_stuck_area && debug_data) {
+      debug_data->stuck_targets.objects.push_back(object);
+      return true;
+    }
+  }
+  return false;
+}
+
+static std::string getTurnDirection(lanelet::ConstLanelet lane)
+{
+  return lane.attributeOr("turn_direction", "else");
+}
+
+IntersectionLanelets IntersectionModule::getObjectiveLanelets(
+  lanelet::LaneletMapConstPtr lanelet_map_ptr, lanelet::routing::RoutingGraphPtr routing_graph_ptr,
+  const lanelet::ConstLanelet assigned_lanelet, const lanelet::ConstLanelets & lanelets_on_path,
+  const std::set<lanelet::Id> & associative_ids, const double detection_area_length,
+  const double occlusion_detection_area_length, const bool consider_wrong_direction_vehicle)
+{
+  const auto turn_direction = assigned_lanelet.attributeOr("turn_direction", "else");
+
+  // retrieve a stopline associated with a traffic light
+  bool has_traffic_light = false;
+  if (const auto tl_reg_elems = assigned_lanelet.regulatoryElementsAs<lanelet::TrafficLight>();
+      tl_reg_elems.size() != 0) {
+    const auto tl_reg_elem = tl_reg_elems.front();
+    const auto stopline_opt = tl_reg_elem->stopLine();
+    if (!!stopline_opt) has_traffic_light = true;
+  }
+
+  // for low priority lane
+  // If ego_lane has right of way (i.e. is high priority),
+  // ignore yieldLanelets (i.e. low priority lanes)
+  lanelet::ConstLanelets yield_lanelets{};
+  const auto right_of_ways = assigned_lanelet.regulatoryElementsAs<lanelet::RightOfWay>();
+  for (const auto & right_of_way : right_of_ways) {
+    if (lanelet::utils::contains(right_of_way->rightOfWayLanelets(), assigned_lanelet)) {
+      for (const auto & yield_lanelet : right_of_way->yieldLanelets()) {
+        yield_lanelets.push_back(yield_lanelet);
+        for (const auto & previous_lanelet : routing_graph_ptr->previous(yield_lanelet)) {
+          yield_lanelets.push_back(previous_lanelet);
+        }
+      }
+    }
+  }
+
+  // get all following lanes of previous lane
+  lanelet::ConstLanelets ego_lanelets = lanelets_on_path;
+  for (const auto & previous_lanelet : routing_graph_ptr->previous(assigned_lanelet)) {
+    ego_lanelets.push_back(previous_lanelet);
+    for (const auto & following_lanelet : routing_graph_ptr->following(previous_lanelet)) {
+      if (lanelet::utils::contains(ego_lanelets, following_lanelet)) {
+        continue;
+      }
+      ego_lanelets.push_back(following_lanelet);
+    }
+  }
+
+  // get conflicting lanes on assigned lanelet
+  const auto & conflicting_lanelets =
+    lanelet::utils::getConflictingLanelets(routing_graph_ptr, assigned_lanelet);
+  std::vector<lanelet::ConstLanelet> adjacent_followings;
+
+  for (const auto & conflicting_lanelet : conflicting_lanelets) {
+    for (const auto & following_lanelet : routing_graph_ptr->following(conflicting_lanelet)) {
+      adjacent_followings.push_back(following_lanelet);
+    }
+    for (const auto & following_lanelet : routing_graph_ptr->previous(conflicting_lanelet)) {
+      adjacent_followings.push_back(following_lanelet);
+    }
+  }
+
+  // final objective lanelets
+  lanelet::ConstLanelets detection_lanelets;
+  lanelet::ConstLanelets conflicting_ex_ego_lanelets;
+  // conflicting lanes is necessary to get stopline for stuck vehicle
+  for (auto && conflicting_lanelet : conflicting_lanelets) {
+    if (!lanelet::utils::contains(ego_lanelets, conflicting_lanelet))
+      conflicting_ex_ego_lanelets.push_back(conflicting_lanelet);
+  }
+
+  // exclude yield lanelets and ego lanelets from detection_lanelets
+  if (turn_direction == std::string("straight") && has_traffic_light) {
+    // if assigned lanelet is "straight" with traffic light, detection area is not necessary
+  } else {
+    if (consider_wrong_direction_vehicle) {
+      for (const auto & conflicting_lanelet : conflicting_lanelets) {
+        if (lanelet::utils::contains(yield_lanelets, conflicting_lanelet)) {
+          continue;
+        }
+        detection_lanelets.push_back(conflicting_lanelet);
+      }
+      for (const auto & adjacent_following : adjacent_followings) {
+        detection_lanelets.push_back(adjacent_following);
+      }
+    } else {
+      // otherwise we need to know the priority from RightOfWay
+      for (const auto & conflicting_lanelet : conflicting_lanelets) {
+        if (
+          lanelet::utils::contains(yield_lanelets, conflicting_lanelet) ||
+          lanelet::utils::contains(ego_lanelets, conflicting_lanelet)) {
+          continue;
+        }
+        detection_lanelets.push_back(conflicting_lanelet);
+      }
+    }
+  }
+
+  // get possible lanelet path that reaches conflicting_lane longer than given length
+  lanelet::ConstLanelets detection_and_preceding_lanelets;
+  {
+    const double length = detection_area_length;
+    std::set<lanelet::Id> detection_ids;
+    for (const auto & ll : detection_lanelets) {
+      // Preceding lanes does not include detection_lane so add them at the end
+      const auto & inserted = detection_ids.insert(ll.id());
+      if (inserted.second) detection_and_preceding_lanelets.push_back(ll);
+      // get preceding lanelets without ego_lanelets
+      // to prevent the detection area from including the ego lanes and its' preceding lanes.
+      const auto lanelet_sequences = lanelet::utils::query::getPrecedingLaneletSequences(
+        routing_graph_ptr, ll, length, ego_lanelets);
+      for (const auto & ls : lanelet_sequences) {
+        for (const auto & l : ls) {
+          const auto & inserted = detection_ids.insert(l.id());
+          if (inserted.second) detection_and_preceding_lanelets.push_back(l);
+        }
+      }
+    }
+  }
+
+  lanelet::ConstLanelets occlusion_detection_and_preceding_lanelets;
+  {
+    const double length = occlusion_detection_area_length;
+    std::set<lanelet::Id> detection_ids;
+    for (const auto & ll : detection_lanelets) {
+      // Preceding lanes does not include detection_lane so add them at the end
+      const auto & inserted = detection_ids.insert(ll.id());
+      if (inserted.second) occlusion_detection_and_preceding_lanelets.push_back(ll);
+      // get preceding lanelets without ego_lanelets
+      // to prevent the detection area from including the ego lanes and its' preceding lanes.
+      const auto lanelet_sequences = lanelet::utils::query::getPrecedingLaneletSequences(
+        routing_graph_ptr, ll, length, ego_lanelets);
+      for (const auto & ls : lanelet_sequences) {
+        for (const auto & l : ls) {
+          const auto & inserted = detection_ids.insert(l.id());
+          if (inserted.second) occlusion_detection_and_preceding_lanelets.push_back(l);
+        }
+      }
+    }
+  }
+  lanelet::ConstLanelets occlusion_detection_and_preceding_lanelets_wo_turn_direction;
+  for (const auto & ll : occlusion_detection_and_preceding_lanelets) {
+    const auto turn_direction = getTurnDirection(ll);
+    if (turn_direction == "left" || turn_direction == "right") {
+      continue;
+    }
+    occlusion_detection_and_preceding_lanelets_wo_turn_direction.push_back(ll);
+  }
+
+  auto [attention_lanelets, original_attention_lanelet_sequences] =
+    util::mergeLaneletsByTopologicalSort(detection_and_preceding_lanelets, routing_graph_ptr);
+
+  IntersectionLanelets result;
+  result.attention_ = std::move(attention_lanelets);
+  for (const auto & original_attention_lanelet_seq : original_attention_lanelet_sequences) {
+    // NOTE: in mergeLaneletsByTopologicalSort(), sub_ids are empty checked, so it is ensured that
+    // back() exists.
+    std::optional<lanelet::ConstLineString3d> stopline{std::nullopt};
+    for (auto it = original_attention_lanelet_seq.rbegin();
+         it != original_attention_lanelet_seq.rend(); ++it) {
+      const auto traffic_lights = it->regulatoryElementsAs<lanelet::TrafficLight>();
+      for (const auto & traffic_light : traffic_lights) {
+        const auto stopline_opt = traffic_light->stopLine();
+        if (!stopline_opt) continue;
+        stopline = stopline_opt.get();
+        break;
+      }
+      if (stopline) break;
+    }
+    result.attention_stoplines_.push_back(stopline);
+  }
+  result.attention_non_preceding_ = std::move(detection_lanelets);
+  for (unsigned i = 0; i < result.attention_non_preceding_.size(); ++i) {
+    std::optional<lanelet::ConstLineString3d> stopline = std::nullopt;
+    const auto & ll = result.attention_non_preceding_.at(i);
+    const auto traffic_lights = ll.regulatoryElementsAs<lanelet::TrafficLight>();
+    for (const auto & traffic_light : traffic_lights) {
+      const auto stopline_opt = traffic_light->stopLine();
+      if (!stopline_opt) continue;
+      stopline = stopline_opt.get();
+    }
+    result.attention_non_preceding_stoplines_.push_back(stopline);
+  }
+  result.conflicting_ = std::move(conflicting_ex_ego_lanelets);
+  result.adjacent_ = planning_utils::getConstLaneletsFromIds(lanelet_map_ptr, associative_ids);
+  // NOTE: occlusion_attention is not inverted here
+  // TODO(Mamoru Sobue): apply mergeLaneletsByTopologicalSort for occlusion lanelets as well and
+  // then trim part of them based on curvature threshold
+  result.occlusion_attention_ =
+    std::move(occlusion_detection_and_preceding_lanelets_wo_turn_direction);
+
+  // NOTE: to properly update(), each element in conflicting_/conflicting_area_,
+  // attention_non_preceding_/attention_non_preceding_area_ need to be matched
+  result.attention_area_ = getPolygon3dFromLanelets(result.attention_);
+  result.attention_non_preceding_area_ = getPolygon3dFromLanelets(result.attention_non_preceding_);
+  result.conflicting_area_ = getPolygon3dFromLanelets(result.conflicting_);
+  result.adjacent_area_ = getPolygon3dFromLanelets(result.adjacent_);
+  result.occlusion_attention_area_ = getPolygon3dFromLanelets(result.occlusion_attention_);
+  return result;
+}
+
+std::optional<IntersectionStopLines> IntersectionModule::generateIntersectionStopLines(
+  const lanelet::CompoundPolygon3d & first_conflicting_area,
+  const lanelet::CompoundPolygon3d & first_attention_area,
+  const lanelet::ConstLineString2d & first_attention_lane_centerline,
+  const std::shared_ptr<const PlannerData> & planner_data,
+  const util::InterpolatedPathInfo & interpolated_path_info, const bool use_stuck_stopline,
+  const double stopline_margin, const double max_accel, const double max_jerk,
+  const double delay_response_time, const double peeking_offset,
+  autoware_auto_planning_msgs::msg::PathWithLaneId * original_path)
+{
+  const auto & path_ip = interpolated_path_info.path;
+  const double ds = interpolated_path_info.ds;
+  const auto & lane_interval_ip = interpolated_path_info.lane_id_interval.value();
+  const double baselink2front = planner_data->vehicle_info_.max_longitudinal_offset_m;
+
+  const int stopline_margin_idx_dist = std::ceil(stopline_margin / ds);
+  const int base2front_idx_dist =
+    std::ceil(planner_data->vehicle_info_.max_longitudinal_offset_m / ds);
+
+  // find the index of the first point whose vehicle footprint on it intersects with detection_area
+  const auto local_footprint = planner_data->vehicle_info_.createFootprint(0.0, 0.0);
+  std::optional<size_t> first_footprint_inside_detection_ip_opt =
+    getFirstPointInsidePolygonByFootprint(
+      first_attention_area, interpolated_path_info, local_footprint, baselink2front);
+  if (!first_footprint_inside_detection_ip_opt) {
+    return std::nullopt;
+  }
+  const auto first_footprint_inside_detection_ip = first_footprint_inside_detection_ip_opt.value();
+
+  std::optional<size_t> first_footprint_attention_centerline_ip_opt = std::nullopt;
+  for (auto i = std::get<0>(lane_interval_ip); i < std::get<1>(lane_interval_ip); ++i) {
+    const auto & base_pose = path_ip.points.at(i).point.pose;
+    const auto path_footprint = tier4_autoware_utils::transformVector(
+      local_footprint, tier4_autoware_utils::pose2transform(base_pose));
+    if (bg::intersects(path_footprint, first_attention_lane_centerline.basicLineString())) {
+      // NOTE: maybe consideration of braking dist is necessary
+      first_footprint_attention_centerline_ip_opt = i;
+      break;
+    }
+  }
+  if (!first_footprint_attention_centerline_ip_opt) {
+    return std::nullopt;
+  }
+  const size_t first_footprint_attention_centerline_ip =
+    first_footprint_attention_centerline_ip_opt.value();
+
+  // (1) default stop line position on interpolated path
+  bool default_stopline_valid = true;
+  int stop_idx_ip_int = -1;
+  if (const auto map_stop_idx_ip = getStopLineIndexFromMap(interpolated_path_info, planner_data);
+      map_stop_idx_ip) {
+    stop_idx_ip_int = static_cast<int>(map_stop_idx_ip.value()) - base2front_idx_dist;
+  }
+  if (stop_idx_ip_int < 0) {
+    stop_idx_ip_int = first_footprint_inside_detection_ip - stopline_margin_idx_dist;
+  }
+  if (stop_idx_ip_int < 0) {
+    default_stopline_valid = false;
+  }
+  const auto default_stopline_ip = stop_idx_ip_int >= 0 ? static_cast<size_t>(stop_idx_ip_int) : 0;
+
+  // (2) ego front stop line position on interpolated path
+  const geometry_msgs::msg::Pose & current_pose = planner_data->current_odometry->pose;
+  const auto closest_idx_ip = motion_utils::findFirstNearestIndexWithSoftConstraints(
+    path_ip.points, current_pose, planner_data->ego_nearest_dist_threshold,
+    planner_data->ego_nearest_yaw_threshold);
+
+  // (3) occlusion peeking stop line position on interpolated path
+  int occlusion_peeking_line_ip_int = static_cast<int>(default_stopline_ip);
+  bool occlusion_peeking_line_valid = true;
+  // NOTE: if footprints[0] is already inside the detection area, invalid
+  {
+    const auto & base_pose0 = path_ip.points.at(default_stopline_ip).point.pose;
+    const auto path_footprint0 = tier4_autoware_utils::transformVector(
+      local_footprint, tier4_autoware_utils::pose2transform(base_pose0));
+    if (bg::intersects(
+          path_footprint0, lanelet::utils::to2D(first_attention_area).basicPolygon())) {
+      occlusion_peeking_line_valid = false;
+    }
+  }
+  if (occlusion_peeking_line_valid) {
+    occlusion_peeking_line_ip_int =
+      first_footprint_inside_detection_ip + std::ceil(peeking_offset / ds);
+  }
+
+  const auto occlusion_peeking_line_ip = static_cast<size_t>(
+    std::clamp<int>(occlusion_peeking_line_ip_int, 0, static_cast<int>(path_ip.points.size()) - 1));
+  const auto first_attention_stopline_ip = first_footprint_inside_detection_ip;
+  const bool first_attention_stopline_valid = true;
+
+  // (4) pass judge line position on interpolated path
+  const double velocity = planner_data->current_velocity->twist.linear.x;
+  const double acceleration = planner_data->current_acceleration->accel.accel.linear.x;
+  const double braking_dist = planning_utils::calcJudgeLineDistWithJerkLimit(
+    velocity, acceleration, max_accel, max_jerk, delay_response_time);
+  int pass_judge_ip_int =
+    static_cast<int>(first_footprint_inside_detection_ip) - std::ceil(braking_dist / ds);
+  const auto pass_judge_line_ip = static_cast<size_t>(
+    std::clamp<int>(pass_judge_ip_int, 0, static_cast<int>(path_ip.points.size()) - 1));
+  // TODO(Mamoru Sobue): maybe braking dist should be considered
+  const auto occlusion_wo_tl_pass_judge_line_ip =
+    static_cast<size_t>(first_footprint_attention_centerline_ip);
+
+  // (5) stuck vehicle stop line
+  int stuck_stopline_ip_int = 0;
+  bool stuck_stopline_valid = true;
+  if (use_stuck_stopline) {
+    // NOTE: when ego vehicle is approaching detection area and already passed
+    // first_conflicting_area, this could be null.
+    const auto stuck_stopline_idx_ip_opt = getFirstPointInsidePolygonByFootprint(
+      first_conflicting_area, interpolated_path_info, local_footprint, baselink2front);
+    if (!stuck_stopline_idx_ip_opt) {
+      stuck_stopline_valid = false;
+      stuck_stopline_ip_int = 0;
+    } else {
+      stuck_stopline_ip_int = stuck_stopline_idx_ip_opt.value() - stopline_margin_idx_dist;
+    }
+  } else {
+    stuck_stopline_ip_int =
+      std::get<0>(lane_interval_ip) - (stopline_margin_idx_dist + base2front_idx_dist);
+  }
+  if (stuck_stopline_ip_int < 0) {
+    stuck_stopline_valid = false;
+  }
+  const auto stuck_stopline_ip = static_cast<size_t>(std::max(0, stuck_stopline_ip_int));
+
+  struct IntersectionStopLinesTemp
+  {
+    size_t closest_idx{0};
+    size_t stuck_stopline{0};
+    size_t default_stopline{0};
+    size_t first_attention_stopline{0};
+    size_t occlusion_peeking_stopline{0};
+    size_t pass_judge_line{0};
+    size_t occlusion_wo_tl_pass_judge_line{0};
+  };
+
+  IntersectionStopLinesTemp intersection_stoplines_temp;
+  std::list<std::pair<const size_t *, size_t *>> stoplines = {
+    {&closest_idx_ip, &intersection_stoplines_temp.closest_idx},
+    {&stuck_stopline_ip, &intersection_stoplines_temp.stuck_stopline},
+    {&default_stopline_ip, &intersection_stoplines_temp.default_stopline},
+    {&first_attention_stopline_ip, &intersection_stoplines_temp.first_attention_stopline},
+    {&occlusion_peeking_line_ip, &intersection_stoplines_temp.occlusion_peeking_stopline},
+    {&pass_judge_line_ip, &intersection_stoplines_temp.pass_judge_line},
+    {&occlusion_wo_tl_pass_judge_line_ip,
+     &intersection_stoplines_temp.occlusion_wo_tl_pass_judge_line}};
+  stoplines.sort(
+    [](const auto & it1, const auto & it2) { return *(std::get<0>(it1)) < *(std::get<0>(it2)); });
+  for (const auto & [stop_idx_ip, stop_idx] : stoplines) {
+    const auto & insert_point = path_ip.points.at(*stop_idx_ip).point.pose;
+    const auto insert_idx = util::insertPointIndex(
+      insert_point, original_path, planner_data->ego_nearest_dist_threshold,
+      planner_data->ego_nearest_yaw_threshold);
+    if (!insert_idx) {
+      return std::nullopt;
+    }
+    *stop_idx = insert_idx.value();
+  }
+  if (
+    intersection_stoplines_temp.occlusion_peeking_stopline <
+    intersection_stoplines_temp.default_stopline) {
+    intersection_stoplines_temp.occlusion_peeking_stopline =
+      intersection_stoplines_temp.default_stopline;
+  }
+
+  IntersectionStopLines intersection_stoplines;
+  intersection_stoplines.closest_idx = intersection_stoplines_temp.closest_idx;
+  if (stuck_stopline_valid) {
+    intersection_stoplines.stuck_stopline = intersection_stoplines_temp.stuck_stopline;
+  }
+  if (default_stopline_valid) {
+    intersection_stoplines.default_stopline = intersection_stoplines_temp.default_stopline;
+  }
+  if (first_attention_stopline_valid) {
+    intersection_stoplines.first_attention_stopline =
+      intersection_stoplines_temp.first_attention_stopline;
+  }
+  if (occlusion_peeking_line_valid) {
+    intersection_stoplines.occlusion_peeking_stopline =
+      intersection_stoplines_temp.occlusion_peeking_stopline;
+  }
+  intersection_stoplines.pass_judge_line = intersection_stoplines_temp.pass_judge_line;
+  intersection_stoplines.occlusion_wo_tl_pass_judge_line =
+    intersection_stoplines_temp.occlusion_wo_tl_pass_judge_line;
+  return intersection_stoplines;
+}
+
+TrafficPrioritizedLevel IntersectionModule::getTrafficPrioritizedLevel(
+  lanelet::ConstLanelet lane, const std::map<int, TrafficSignalStamped> & tl_infos)
+{
+  using TrafficSignalElement = autoware_perception_msgs::msg::TrafficSignalElement;
+
+  std::optional<int> tl_id = std::nullopt;
+  for (auto && tl_reg_elem : lane.regulatoryElementsAs<lanelet::TrafficLight>()) {
+    tl_id = tl_reg_elem->id();
+    break;
+  }
+  if (!tl_id) {
+    // this lane has no traffic light
+    return TrafficPrioritizedLevel::NOT_PRIORITIZED;
+  }
+  const auto tl_info_it = tl_infos.find(tl_id.value());
+  if (tl_info_it == tl_infos.end()) {
+    // the info of this traffic light is not available
+    return TrafficPrioritizedLevel::NOT_PRIORITIZED;
+  }
+  const auto & tl_info = tl_info_it->second;
+  bool has_amber_signal{false};
+  for (auto && tl_light : tl_info.signal.elements) {
+    if (tl_light.color == TrafficSignalElement::AMBER) {
+      has_amber_signal = true;
+    }
+    if (tl_light.color == TrafficSignalElement::RED) {
+      // NOTE: Return here since the red signal has the highest priority.
+      return TrafficPrioritizedLevel::FULLY_PRIORITIZED;
+    }
+  }
+  if (has_amber_signal) {
+    return TrafficPrioritizedLevel::PARTIALLY_PRIORITIZED;
+  }
+  return TrafficPrioritizedLevel::NOT_PRIORITIZED;
+}
+
+std::optional<size_t> IntersectionModule::generateStuckStopLine(
+  const lanelet::CompoundPolygon3d & conflicting_area,
+  const std::shared_ptr<const PlannerData> & planner_data,
+  const util::InterpolatedPathInfo & interpolated_path_info, const double stopline_margin,
+  const bool use_stuck_stopline, autoware_auto_planning_msgs::msg::PathWithLaneId * original_path)
+{
+  const auto & path_ip = interpolated_path_info.path;
+  const double ds = interpolated_path_info.ds;
+  const auto & lane_interval_ip = interpolated_path_info.lane_id_interval.value();
+  const auto lane_interval_ip_start = std::get<0>(lane_interval_ip);
+  size_t stuck_stopline_idx_ip = 0;
+  if (use_stuck_stopline) {
+    stuck_stopline_idx_ip = lane_interval_ip_start;
+  } else {
+    const auto stuck_stopline_idx_ip_opt =
+      util::getFirstPointInsidePolygon(path_ip, lane_interval_ip, conflicting_area);
+    if (!stuck_stopline_idx_ip_opt) {
+      return std::nullopt;
+    }
+    stuck_stopline_idx_ip = stuck_stopline_idx_ip_opt.value();
+  }
+
+  const int stopline_margin_idx_dist = std::ceil(stopline_margin / ds);
+  const int base2front_idx_dist =
+    std::ceil(planner_data->vehicle_info_.max_longitudinal_offset_m / ds);
+  const size_t insert_idx_ip = static_cast<size_t>(std::max(
+    static_cast<int>(stuck_stopline_idx_ip) - 1 - stopline_margin_idx_dist - base2front_idx_dist,
+    0));
+  const auto & insert_point = path_ip.points.at(insert_idx_ip).point.pose;
+  return util::insertPointIndex(
+    insert_point, original_path, planner_data->ego_nearest_dist_threshold,
+    planner_data->ego_nearest_yaw_threshold);
+}
+
+/**
+ * @brief Get stop point from map if exists
+ * @param stop_pose stop point defined on map
+ * @return true when the stop point is defined on map.
+ */
+std::optional<size_t> IntersectionModule::getStopLineIndexFromMap(
+  const util::InterpolatedPathInfo & interpolated_path_info,
+  const std::shared_ptr<const PlannerData> & planner_data)
+{
+  const auto & path = interpolated_path_info.path;
+  const auto & lane_interval = interpolated_path_info.lane_id_interval.value();
+
+  lanelet::ConstLanelet lanelet =
+    planner_data->route_handler_->getLaneletMapPtr()->laneletLayer.get(
+      interpolated_path_info.lane_id);
+  const auto road_markings = lanelet.regulatoryElementsAs<lanelet::autoware::RoadMarking>();
+  lanelet::ConstLineStrings3d stopline;
+  for (const auto & road_marking : road_markings) {
+    const std::string type =
+      road_marking->roadMarking().attributeOr(lanelet::AttributeName::Type, "none");
+    if (type == lanelet::AttributeValueString::StopLine) {
+      stopline.push_back(road_marking->roadMarking());
+      break;  // only one stopline exists.
+    }
+  }
+  if (stopline.empty()) {
+    return std::nullopt;
+  }
+
+  const auto p_start = stopline.front().front();
+  const auto p_end = stopline.front().back();
+  const LineString2d extended_stopline =
+    planning_utils::extendLine(p_start, p_end, planner_data->stop_line_extend_length);
+
+  for (size_t i = lane_interval.first; i < lane_interval.second; i++) {
+    const auto & p_front = path.points.at(i).point.pose.position;
+    const auto & p_back = path.points.at(i + 1).point.pose.position;
+
+    const LineString2d path_segment = {{p_front.x, p_front.y}, {p_back.x, p_back.y}};
+    std::vector<Point2d> collision_points;
+    bg::intersection(extended_stopline, path_segment, collision_points);
+
+    if (collision_points.empty()) {
+      continue;
+    }
+
+    return i;
+  }
+
+  geometry_msgs::msg::Pose stop_point_from_map;
+  stop_point_from_map.position.x = 0.5 * (p_start.x() + p_end.x());
+  stop_point_from_map.position.y = 0.5 * (p_start.y() + p_end.y());
+  stop_point_from_map.position.z = 0.5 * (p_start.z() + p_end.z());
+
+  return motion_utils::findFirstNearestIndexWithSoftConstraints(
+    path.points, stop_point_from_map, planner_data->ego_nearest_dist_threshold,
+    planner_data->ego_nearest_yaw_threshold);
 }
 
 /*
